feat: added theme schemas to common & updated imports & bump (#2309)
* feat: added theme schemas to common & updated imports & bump

* fix: lock fix

* chore(stam): trigger workflow

* Update ci.yml

---------
