--- apps/backoffice-v2/src/domains/workflow-definitions/enums/workflow-definition-config-theme.ts
@@ -1,11 +1,0 @@
-export const WorkflowDefinitionConfigThemeEnum = {
-  KYC: 'kyc',
-  KYB: 'kyb',
-  DOCUMENTS_REVIEW: 'documents-review',
-} as const;
-
-export const WorkflowDefinitionConfigThemes = [
-  WorkflowDefinitionConfigThemeEnum.KYB,
-  WorkflowDefinitionConfigThemeEnum.KYC,
-  WorkflowDefinitionConfigThemeEnum.DOCUMENTS_REVIEW,
-] as const;
