+++ apps/backoffice-v2/src/domains/workflow-definitions/fetchers.ts
@@ -3,12 +3,12 @@ import { Method } from '@/common/enums';
 import { env } from '@/common/env/env';
 import { getOriginUrl } from '@/common/utils/get-origin-url/get-url-origin';
 import { handleZodError } from '@/common/utils/handle-zod-error/handle-zod-error';
+import { ObjectWithIdSchema } from '@/lib/zod/utils/object-with-id/object-with-id';
 import {
   WorkflowDefinitionConfigThemeEnum,
-  WorkflowDefinitionConfigThemes,
-} from '@/domains/workflow-definitions/enums/workflow-definition-config-theme';
-import { ObjectWithIdSchema } from '@/lib/zod/utils/object-with-id/object-with-id';
-import { WorkflowDefinitionVariant } from '@ballerine/common';
+  WorkflowDefinitionConfigThemeSchema,
+  WorkflowDefinitionVariant,
+} from '@ballerine/common';
 import { z } from 'zod';
 
 export const PluginSchema = z.object({
@@ -18,19 +18,14 @@ export const PluginSchema = z.object({
 
 export type TPlugin = z.infer<typeof PluginSchema>;
 
-export const WorkflowDefinitionConfigTheme = z.object({
-  type: z.enum(WorkflowDefinitionConfigThemes).default(WorkflowDefinitionConfigThemeEnum.KYB),
-  tabsOverride: z.array(z.string()).optional(),
-});
-
-export type WorkflowDefinitionConfigTheme = z.infer<typeof WorkflowDefinitionConfigTheme>;
+export type WorkflowDefinitionConfigTheme = z.infer<typeof WorkflowDefinitionConfigThemeSchema>;
 
 export const WorkflowDefinitionConfigSchema = z
   .object({
     enableManualCreation: z.boolean().default(false),
     isManualCreation: z.boolean().default(false),
     isAssociatedCompanyKybEnabled: z.boolean().default(false),
-    theme: WorkflowDefinitionConfigTheme.default({
+    theme: WorkflowDefinitionConfigThemeSchema.default({
       type: WorkflowDefinitionConfigThemeEnum.KYB,
     }),
   })
