+++ apps/kyb-app/CHANGELOG.md
@@ -1,5 +1,13 @@
 # kyb-app
 
+## 0.3.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.3
+  - @ballerine/workflow-browser-sdk@0.6.6
+
 ## 0.3.5
 
 ### Patch Changes
