+++ apps/kyb-app/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@ballerine/kyb-app",
   "private": true,
-  "version": "0.3.5",
+  "version": "0.3.6",
   "type": "module",
   "scripts": {
     "dev": "vite",
@@ -15,9 +15,9 @@
   },
   "dependencies": {
     "@ballerine/blocks": "0.2.2",
-    "@ballerine/common": "^0.9.2",
+    "@ballerine/common": "^0.9.3",
     "@ballerine/ui": "0.5.2",
-    "@ballerine/workflow-browser-sdk": "0.6.5",
+    "@ballerine/workflow-browser-sdk": "0.6.6",
     "@lukemorales/query-key-factory": "^1.0.3",
     "@radix-ui/react-icons": "^1.3.0",
     "@rjsf/core": "^5.9.0",
