+++ packages/common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ballerine/common
 
+## 0.9.3
+
+### Patch Changes
+
+- Added workflow definition theme schemas
+
 ## 0.9.2
 
 ### Patch Changes
