+++ packages/common/package.json
@@ -2,7 +2,7 @@
   "private": false,
   "name": "@ballerine/common",
   "author": "Ballerine <dev@ballerine.com>",
-  "version": "0.9.2",
+  "version": "0.9.3",
   "description": "common",
   "module": "./dist/esm/index.js",
   "main": "./dist/cjs/index.js",
@@ -74,12 +74,12 @@
     "ts-node": "^10.9.1",
     "typescript": "4.9.5",
     "vite": "^4.5.3",
-    "vitest": "^0.28.4",
-    "zod": "^3.22.3"
+    "vitest": "^0.28.4"
   },
   "dependencies": {
     "@sinclair/typebox": "^0.31.7",
     "ajv": "^8.12.0",
-    "json-schema-to-zod": "^0.6.3"
+    "json-schema-to-zod": "^0.6.3",
+    "zod": "^3.22.4"
   }
 }
