+++ packages/common/src/consts/index.ts
@@ -95,3 +95,17 @@ export const UnifiedApiReasons = [
 export type TUnifiedApiReason = (typeof UnifiedApiReasons)[number];
 
 export type TUnifiedApiReasons = typeof UnifiedApiReasons;
+
+export const WorkflowDefinitionConfigThemeEnum = {
+  KYC: 'kyc',
+  KYB: 'kyb',
+  DOCUMENTS_REVIEW: 'documents-review',
+} as const;
+
+export const WorkflowDefinitionConfigThemes = [
+  WorkflowDefinitionConfigThemeEnum.KYB,
+  WorkflowDefinitionConfigThemeEnum.KYC,
+  WorkflowDefinitionConfigThemeEnum.DOCUMENTS_REVIEW,
+] as const;
+
+export type TWorkflowDefinitionConfigTheme = (typeof WorkflowDefinitionConfigThemes)[number];
