+++ packages/common/src/schemas/workflow/workflow-config-theme.ts
@@ -1,0 +1,7 @@
+import { WorkflowDefinitionConfigThemeEnum, WorkflowDefinitionConfigThemes } from '@/consts';
+import { z } from 'zod';
+
+export const WorkflowDefinitionConfigThemeSchema = z.object({
+  type: z.enum(WorkflowDefinitionConfigThemes).default(WorkflowDefinitionConfigThemeEnum.KYB),
+  tabsOverride: z.array(z.string()).optional(),
+});
