+++ packages/workflow-core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ballerine/workflow-core
 
+## 0.6.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.3
+
 ## 0.6.5
 
 ### Patch Changes
