+++ sdks/web-ui-sdk/CHANGELOG.md
@@ -1,5 +1,12 @@
 # web-ui-sdk
 
+## 1.5.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.3
+
 ## 1.5.3
 
 ### Patch Changes
