+++ sdks/workflow-browser-sdk/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ballerine/workflow-browser-sdk
 
+## 0.6.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.3
+  - @ballerine/workflow-core@0.6.6
+
 ## 0.6.5
 
 ### Patch Changes
