+++ services/workflows-service/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @ballerine/workflows-service
 
+## 0.7.5
+
+### Patch Changes
+
+- Added workflow definition theme schemas
+- Updated dependencies
+  - @ballerine/common@0.9.3
+  - @ballerine/workflow-core@0.6.6
+  - @ballerine/workflow-node-sdk@0.6.6
+
 ## 0.7.4
 
 ### Patch Changes
