+++ services/workflows-service/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@ballerine/workflows-service",
   "private": false,
-  "version": "0.7.4",
+  "version": "0.7.5",
   "description": "workflow-service",
   "scripts": {
     "spellcheck": "cspell \"*\"",
@@ -45,9 +45,9 @@
     "@aws-sdk/client-s3": "3.347.1",
     "@aws-sdk/lib-storage": "3.347.1",
     "@aws-sdk/s3-request-presigner": "3.347.1",
-    "@ballerine/common": "0.9.2",
-    "@ballerine/workflow-core": "0.6.5",
-    "@ballerine/workflow-node-sdk": "0.6.5",
+    "@ballerine/common": "0.9.3",
+    "@ballerine/workflow-core": "0.6.6",
+    "@ballerine/workflow-node-sdk": "0.6.6",
     "@faker-js/faker": "^7.6.0",
     "@nestjs/axios": "^2.0.0",
     "@nestjs/common": "^9.3.12",
